add RECORDINGS_METRICS_READ_FROM_DISK setting
If the exporter is run on the same host as BigBlueButton, we can
collect expensive recording metrics from the disk rather than from
the API. Because the API response for recording metrics returns an
entire XML object for each entry scraping becomes slow. The
optimization is to simply count the number of files on disk in
the appropriate directory, requiring read only access to the
directory: "/var/bigbluebutton".